refactor(/game): 체스 게임에서 사용되는 객체들을 game 패키지로 모음
