+++ src/test/java/domain/game/PositionTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 import domain.position.File;
 import domain.position.Position;
